+++ Ticketing/EndPoints/Reporting/Query/GetReport/ReportHeaderInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace Ticketing.EndPoints.Reporting.Query.GetReport
+{
+    public class ReportHeaderInfo
+    {
+        public string Title { get; set; }
+        public string startDate { get; set; }
+        public string endDate { get; set; }
+        public string PrintDate { get; set; }
+    }
+}
